// apps/api/src/modules/external-data/infrastructure/queues/external-data-queue.constants.ts
export const EXTERNAL_DATA_QUEUE_NAME = "external-data-ingestion";

export enum ExternalDataJobType {
  POLL_RSS_SOURCES = "POLL_RSS_SOURCES",
  POLL_BINANCE_ANNOUNCEMENTS = "POLL_BINANCE_ANNOUNCEMENTS",
  POLL_OKX_ANNOUNCEMENTS = "POLL_OKX_ANNOUNCEMENTS",
  POLL_FEAR_GREED = "POLL_FEAR_GREED",
  POLL_REDDIT = "POLL_REDDIT",
  POLL_OFFICIAL_MACRO = "POLL_OFFICIAL_MACRO",
  PROVIDER_HEALTH_SCAN = "PROVIDER_HEALTH_SCAN",
  RETENTION_CLEANUP = "RETENTION_CLEANUP",
}